Added support for CMAKE_CXX_FLAGS.
Before this commit, if you put for example: -Wall -Wextra in
CMAKE_CXX_FLAGS - it wouldn't compile, because it interpreted our
addition of -std=c++11 as a list and separated the components with
semicolons, which freaked out the compiler.

Now it treats it as a big string separated with spaces.