move constructor added to curl_easy
A curl_easy object is now movable, that means that the curl handle is
not duped between instances but transfered.

This makes possible to store curl_easy instances in a vector; without
the move constructor, if a resize of the vector happens, the combo (copy
constructor + destructor) will be called cancelling any pending
transfer.